# Conflicts:
#	Client/clientapp/src/Pages/GamesPage.tsx

+++ Server/Services/TransferModels/Responses/GameDto.cs
@@ -31,9 +31,9 @@ public class GameDto
 
     public int? Year { get; set; }
 
-    public decimal? Prizepool { get; set; }
+    public int? Prizepool { get; set; }
 
-    public decimal? Carryover { get; set; }
+    public int? Carryover { get; set; }
 
     public bool? Isactive { get; set; }
